% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyvalidate-package.R
\docType{package}
\name{tidyvalidate-package}
\alias{tidyvalidate}
\alias{tidyvalidate-package}
\title{tidyvalidate: Simplify the use of 'validate' package.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package simplifies the workflow of the 'validate' package to work with data.frames by providing tools to manage the usage of 'confront', 'validator', 'violating', and 'summary' functions. It also streamlines the integration with base R's 'stop' and 'warning' functions, enhancing error handling and reporting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://angelfelizr.github.io/tidyvalidate/}
}

}
\author{
\strong{Maintainer}: Angel Feliz \email{angel.esteban.feliz@gmail.com}

}
\keyword{internal}
