% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_rules.R
\name{validate_rules}
\alias{validate_rules}
\title{Validate domain knowing in a data.frame}
\usage{
validate_rules(
  df,
  ...,
  env = list(),
  keep_rl_cols = NULL,
  select_rl_rules = NULL
)
}
\arguments{
\item{df}{A data.frame or data.table to perform the analysis.}

\item{...}{A comma-separated list of validating expressions}

\item{env}{A list with all the variables we want to make available in the validation rules.}

\item{keep_rl_cols}{Defining the columns to keep all data.table of broken row level rules.}

\item{select_rl_rules}{Defining which row level rules to return as data.table in the row_level_errors element.}
}
\value{
This function return a list of 2 elements:
\itemize{
\item summary: Return data.table with the result of all checks.
\item row_level_errors: Return a list of data.frames containing the column \strong{Broken Rule}, the columns listed in the \code{keep_rl_cols} and the columns used to perform the validation.
}
}
\description{
This function if a wrappers from many utilities from the validate package.
}
\examples{
validation_list <- 
  data.table::as.data.table(mtcars, 
                            keep.rownames = "Car (Name)") |>
  validate_rules("mpg low" = mpg > min_mpg,
                 "hp high" = hp < 200,
                 env = list(min_mpg = 15),
                 keep_rl_cols = "Car (Name)")

names(validation_list)

validation_list$summary

validation_list$row_level_errors
}
