% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_if_problem.R
\name{action_if_problem}
\alias{action_if_problem}
\title{Stop the script if any problem found}
\usage{
action_if_problem(x, message_text, problem_action = c("stop", "warning"))
}
\arguments{
\item{x}{A list containing the summary from \code{\link[=validate_rules]{validate_rules()}}.}

\item{message_text}{The message to display when applying the action.}

\item{problem_action}{Select if you want to display warning or stop the code.}
}
\value{
The list passed as first argument.
}
\description{
Stop the script if any problem found
}
\examples{

try({
  validate_rules(mtcars, "hp high" = hp < 200) |>
    action_if_problem("Here is the error", "stop")
})


validate_rules(mtcars, "hp high" = hp < 200) |>
  action_if_problem("Here is the warning", "warning")
}
