% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyvalidate-package.R
\docType{package}
\name{tidyvalidate-package}
\alias{tidyvalidate}
\alias{tidyvalidate-package}
\title{tidyvalidate: Tidy Data Validation Infrastructure}
\description{
Simplify the workflow needed to use the tools available in the 'validate' package.
}
\author{
\strong{Maintainer}: Angel Feliz \email{angel.esteban.feliz@gmail.com}

}
\keyword{internal}
